omini_text/requirements.txt
# Core dependencies for Omini-Text unified interface
pyyaml>=6.0
python-dotenv>=1.0.0

# For Glimpse detector
openai>=1.0.0
numpy>=1.24.0
scipy>=1.10.0
